Extracted CompilerVersion, interpreted os families
